graphql: Fix an inheritance bug.
Inheriting a link and changing its target type caused an error in
GraphQL.

GraphQL inheritance does not allow changing the field type, so when
reflecting pointers we must use the top-most base pointer version. It is
possible to access the additional fields (if any) of the subclassed
target type if the subtype is references by name:

query {
    Foo {
        bar {
            common
            ... on Bar2 {
                bar2_prop
            }
        }
    }
}

Closes: #709.
